# api/openapi.yaml
openapi: 3.0.3
info:
  title: Bharat Competitive Exam Platform API
  version: 0.1.0
  description: API scaffold for online competitive exams, question bank management, student attempts, and result analytics.
servers:
  - url: https://api.exam-platform.example.com
paths:
  /api/auth/send-otp:
    post:
      tags: [Auth]
      summary: Send OTP for sign-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpRequest'
      responses:
        '200':
          description: OTP accepted for delivery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendOtpResponse'
  /api/auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify OTP and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpRequest'
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
  /api/catalog/categories:
    get:
      tags: [Catalog]
      summary: List exam categories
      responses:
        '200':
          description: Categories returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
  /api/catalog/exams:
    get:
      tags: [Catalog]
      summary: List exams
      parameters:
        - in: query
          name: categoryId
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Exam summaries returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExamSummary'
  /api/catalog/exams/{examId}:
    get:
      tags: [Catalog]
      summary: Get exam detail with test series
      parameters:
        - in: path
          name: examId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Exam detail returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamDetail'
  /api/admin/questions:
    get:
      tags: [Question Bank]
      summary: List question bank entries
      responses:
        '200':
          description: Questions returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
    post:
      tags: [Question Bank]
      summary: Create a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionRequest'
      responses:
        '201':
          description: Question created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
  /api/attempts/start:
    post:
      tags: [Attempts]
      summary: Start a student attempt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartAttemptRequest'
      responses:
        '200':
          description: Attempt created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attempt'
  /api/attempts/{attemptId}:
    get:
      tags: [Attempts]
      summary: Get attempt detail
      parameters:
        - in: path
          name: attemptId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Attempt returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attempt'
  /api/attempts/{attemptId}/answers:
    post:
      tags: [Attempts]
      summary: Save answer during attempt
      parameters:
        - in: path
          name: attemptId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveAnswerRequest'
      responses:
        '200':
          description: Answer saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attempt'
  /api/attempts/{attemptId}/submit:
    post:
      tags: [Attempts]
      summary: Submit attempt and calculate score
      parameters:
        - in: path
          name: attemptId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Attempt submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
  /api/results/{attemptId}:
    get:
      tags: [Results]
      summary: Get result detail
      parameters:
        - in: path
          name: attemptId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Result returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
  /api/admin/overview:
    get:
      tags: [Admin]
      summary: Get admin dashboard overview
      responses:
        '200':
          description: Overview returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOverview'
components:
  schemas:
    MultilingualText:
      type: object
      additionalProperties:
        type: string
    SendOtpRequest:
      type: object
      required: [mobileNumber]
      properties:
        mobileNumber:
          type: string
        role:
          type: string
          enum: [Student, Admin, SuperAdmin]
    SendOtpResponse:
      type: object
      properties:
        mobileNumber:
          type: string
        expiresInSeconds:
          type: integer
        transactionId:
          type: string
    VerifyOtpRequest:
      type: object
      required: [mobileNumber, otp]
      properties:
        mobileNumber:
          type: string
        otp:
          type: string
        role:
          type: string
          enum: [Student, Admin, SuperAdmin]
    Session:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresInSeconds:
          type: integer
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        mobileNumber:
          type: string
        email:
          type: string
        role:
          type: string
        preferredLanguage:
          type: string
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
    ExamSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        totalMarks:
          type: number
        categoryName:
          type: string
        languages:
          type: array
          items:
            type: string
    TestSeries:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        sequenceNumber:
          type: integer
        isFree:
          type: boolean
        releaseDate:
          type: string
          format: date-time
    ExamDetail:
      allOf:
        - $ref: '#/components/schemas/ExamSummary'
        - type: object
          properties:
            negativeMarkingRatio:
              type: number
            testSeries:
              type: array
              items:
                $ref: '#/components/schemas/TestSeries'
    QuestionOption:
      type: object
      properties:
        id:
          type: string
          format: uuid
        text:
          $ref: '#/components/schemas/MultilingualText'
        imageUrl:
          type: string
        isCorrect:
          type: boolean
    Question:
      type: object
      properties:
        id:
          type: string
          format: uuid
        subjectId:
          type: string
          format: uuid
        type:
          type: string
        text:
          $ref: '#/components/schemas/MultilingualText'
        imageUrl:
          type: string
        audioUrl:
          type: string
        explanationText:
          $ref: '#/components/schemas/MultilingualText'
        explanationImageUrl:
          type: string
        marks:
          type: number
        negativeMarks:
          type: number
        difficulty:
          type: string
        tags:
          type: array
          items:
            type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/QuestionOption'
    CreateQuestionRequest:
      type: object
      required: [subjectId, type, text, marks]
      properties:
        subjectId:
          type: string
          format: uuid
        type:
          type: string
        text:
          $ref: '#/components/schemas/MultilingualText'
        imageUrl:
          type: string
        audioUrl:
          type: string
        explanationText:
          $ref: '#/components/schemas/MultilingualText'
        explanationImageUrl:
          type: string
        marks:
          type: number
        negativeMarks:
          type: number
        difficulty:
          type: string
        tags:
          type: array
          items:
            type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/QuestionOption'
    StartAttemptRequest:
      type: object
      required: [studentId, testSeriesId]
      properties:
        studentId:
          type: string
          format: uuid
        testSeriesId:
          type: string
          format: uuid
    SaveAnswerRequest:
      type: object
      required: [questionId]
      properties:
        questionId:
          type: string
          format: uuid
        selectedOptionIds:
          type: array
          items:
            type: string
            format: uuid
        textAnswer:
          type: string
        isMarkedForReview:
          type: boolean
        timeSpentSeconds:
          type: integer
    AttemptQuestion:
      type: object
      properties:
        questionId:
          type: string
          format: uuid
        section:
          type: string
        orderNumber:
          type: integer
        question:
          $ref: '#/components/schemas/Question'
    Attempt:
      type: object
      properties:
        id:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        testSeriesId:
          type: string
          format: uuid
        startedAt:
          type: string
          format: date-time
        submittedAt:
          type: string
          format: date-time
        status:
          type: string
        score:
          type: number
        switchedTabs:
          type: integer
        questions:
          type: array
          items:
            $ref: '#/components/schemas/AttemptQuestion'
    ResultItem:
      type: object
      properties:
        questionId:
          type: string
          format: uuid
        correct:
          type: boolean
        marksAwarded:
          type: number
        explanationText:
          $ref: '#/components/schemas/MultilingualText'
    Result:
      type: object
      properties:
        attemptId:
          type: string
          format: uuid
        score:
          type: number
        totalMarks:
          type: number
        correctAnswers:
          type: integer
        incorrectAnswers:
          type: integer
        accuracyPercentage:
          type: number
        rank:
          type: integer
        percentile:
          type: number
        items:
          type: array
          items:
            $ref: '#/components/schemas/ResultItem'
    AdminOverview:
      type: object
      properties:
        totalExams:
          type: integer
        totalQuestions:
          type: integer
        totalAttempts:
          type: integer
        completedAttempts:
          type: integer
        supportedLanguages:
          type: array
          items:
            type: string